requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.10
fastapi==0.104.1
Flask==3.1.0
gensim==4.3.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
murmurhash==1.0.11
networkx==3.4.2
numpy==1.26.2
packaging==24.2
preshed==3.0.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
requests==2.32.3
scikit-fuzzy==0.4.2
scikit-learn==1.3.2
scipy==1.14.1
setuptools==75.6.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
thinc==8.2.5
threadpoolctl==3.5.0
tqdm==4.67.0
typer==0.9.4
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.24.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.1.3
